dissolve Homepage into components
